# outbound-auth-hmac-demo/client/src/main/resources/openapi/openapi-spec.yml
openapi: 3.0.3
info:
  title: Notify Client API
  description: |
    Client that receives HMAC-authenticated notifications from the server.
    - **POST /subscribe** – client calls server to register; server returns keyId and secret; client stores in local store only.
    - **GET /secret** – get secret from client store only (no server call). Use rotate-secret if lost.
    - **POST /rotate-secret** – client calls server to rotate; client stores new secret for future notifications.
    - **POST /callback-notify** – called by the server with the notification payload and HMAC headers.
  version: 1.0.0
paths:
...
  /subscribe:
    post:
      summary: Subscribe for notifications
      description: |
        Client calls server to register (name, callbackUrl). Server generates keyId and secret,
        returns them in the response; client stores in local store only. Server will POST to
        /callback-notify with HMAC-signed notifications.
      operationId: subscribe
      tags:
        - subscribe
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Subscriber name (included in server notification payload)
      responses:
        '202':
          description: Accepted; returns keyId and secret; client stores them; server will call /callback-notify
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretResponse'
...
components:
  schemas:
    SecretResponse:
      type: object
      properties:
        keyId:
          type: string
        secret:
          type: string